Set limit for separate server mode to 21
In separate server mode, the Max character count for Discord roles is 100. So this is where character count is important unlike the non separate mode where it uses mentions so only IDs take up the character count. But pretty sure no one is gonna max that limit. I guess the average max amount of characters for a role is around 40, and bot can handle up to 1024 characters. But at the same time, need to make space for other characters in the embed so I decided to allow up to 20 roles that can be displayed, up to around 800 characters so there is enough room for the other characters in the embed, including the full username. ModMail devs can adjust this amount if needed.